Documentation done and generated the final JavaDoc
I believe that most of the documentation for this project is done, and
have generated its corresponding JavaDoc; however, there's a bug with
GameLoader when running on other UNIX machines like my MacBook Pro with
a one liner, an have commented on it's respective method. I will bring
this issue up with Dr. Speight and my TA, Ayush.